Refactor(Language selection): Significantly reduced code needed to select language
With help from the Engineering Man Discord, I was able to significantly reduce the number of lines needed to select the project language.
The code now loops through a list of extensions, and checks whether a file exists with that extension.

Although the number of lines has been reduced, there may be a potential bug - not fully tested yet - where if a project file is in a subdirectory. An example would be a Visual Studio project.
(e.g. projectName/projectName/projectFile.cp)