Bug fix: Fixed a bug that appeared with refactored code from last commit
As noted in the previous commit, there was a bug/issue with the new implementation of the code. This meant that any project files that were in a subdirectory could not  be identified, eventually leading to an infinite loop.
The infinite loop was created because the program did not understand what langauge the project was. This would cause an "unknown command" message to loop infinitely.

The issue above has been fixed, but I still need a method to identify web extensions