fix: Fixed infinite loop error
This issue had been partially fixed in a previous commit - #6 - it has now been solved completely, based on testing.
The infinite loop would occure when you tried to backup a file that had been backed up on the same day -- file already exists. The issue was solved by adding a check - before copying and moving - to see if that file exists.

Closes #7