Refactor
Got rid of some old left over code, that was not beimg used. It is used to break out of the infinite - intentional - while loop, but needed to be moved further up and put inside the if statement, instead of it's old location -- from one of the first versions of the program.